+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased changes
+- Added
+  - New GRPC-endpoints: `GetBlocks`, `GetFinalizedBlocks`, `GetBranches`, `GetAncestors`, `GetBlockPendingUpdates`
 
 ## 4.1.0
 - Bugfix